+++ flowpaths/kminpatherror.py
@@ -1,5 +1,5 @@
 import networkx as nx
-import flowpaths.stdigraph as stdigraph
+import flowpaths.stdag as stdag
 import flowpaths.abstractpathmodeldag as pathmodel
 import flowpaths.utils as utils
 import flowpaths.nodeexpandeddigraph as nedg
@@ -198,7 +198,7 @@ class kMinPathError(pathmodel.AbstractPathModelDAG):
             utils.logger.error(f"flow_attr_origin must be either 'node' or 'edge', not {self.flow_attr_origin}")
             raise ValueError(f"flow_attr_origin must be either 'node' or 'edge', not {self.flow_attr_origin}")
 
-        self.G = stdigraph.stDiGraph(self.G_internal, additional_starts=additional_starts_internal, additional_ends=additional_ends_internal)
+        self.G = stdag.stDAG(self.G_internal, additional_starts=additional_starts_internal, additional_ends=additional_ends_internal)
         self.subpath_constraints = subpath_constraints_internal
         self.edges_to_ignore = self.G.source_sink_edges.union(edges_to_ignore_internal)
         self.edge_error_scaling = error_scaling_internal
